Bug 1759324 - update libvpx to v1.11.0 r=bryce
Differential Revision: https://phabricator.services.mozilla.com/D140925

UltraBlame original commit: 0f7ab2324c82f02e9cf9966cb0dde7634d4a3b33

+++ media/libvpx/libvpx/vp8/encoder/treewriter.h
@@ -14,6 +14,8 @@
 
 
 
+#include <stdint.h>
+
 #include "./vpx_config.h"
 #include "vp8/common/treecoder.h"
 
@@ -48,7 +50,9 @@ static INLINE unsigned int vp8_cost_branch(const unsigned int ct[2],
                                            vp8_prob p) {
   
 
-  return ((ct[0] * vp8_cost_zero(p)) + (ct[1] * vp8_cost_one(p))) >> 8;
+  return (unsigned int)(((((uint64_t)ct[0]) * vp8_cost_zero(p)) +
+                         (((uint64_t)ct[1]) * vp8_cost_one(p))) >>
+                        8);
 }
